+++ scim2_models/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseModel
+from .base import CaseExact
 from .base import ComplexAttribute
 from .base import Context
 from .base import Mutability
@@ -56,6 +57,7 @@ __all__ = [
     "BulkOperation",
     "BulkRequest",
     "BulkResponse",
+    "CaseExact",
     "ChangePassword",
     "ComplexAttribute",
     "Context",