+++ doc/changelog.rst
@@ -1,9 +1,12 @@
 Changelog
 =========
 
+Added
+^^^^^
+- Implement :class:`~scim2_models.CaseExact` model annotations.
+
 Changed
 ^^^^^^^
-
 - Refactor :code:`get_field_mutability` and :code:`get_field_returability` in :code:`get_field_annotation`.
 
 [0.1.5] - 2024-06-04